#459: Moved and renamed the file containing the PassiveCoverParser
 The Parser was not modified.
