+++ src/app/ffbe/mappers/effects/passive-effect-parser.factory.ts
@@ -1,7 +1,7 @@
 import {EffectParser} from './effect-parser';
 import {UnknownEffectParser} from './unknown-effect-parser';
 import {PassiveStatsIncreaseHpThresholdParser} from './passives/passive-stats-increase-hp-threshold.parser';
-import {PassiveCoverParser} from './passives/passive-cover.parser';
+import {PassiveCoverParser} from '../../model/effects/passives/passive-cover-effect.model';
 import {PassiveMpRecoveryParser} from './passives/passive-mp-recovery.parser';
 import {PassiveEvasionPhysicalParser} from './passives/passive-evasion-physical.parser';
 import {PassiveTargetChanceChangesParser} from './passives/passive-target-chance-changes.parser';
